input: ensure sessionOpts is properly threaded through
