+++ lnwallet/revocation_producer.go
@@ -3,6 +3,7 @@
 package lnwallet
 
 import (
+	"github.com/lightningnetwork/lnd/channeldb"
 	"github.com/lightningnetwork/lnd/keychain"
 	"github.com/lightningnetwork/lnd/shachain"
 )
@@ -50,7 +51,7 @@ func (l *LightningWallet) nextRevocationProducer(res *ChannelReservation,
 	// Once we have the root, we can then generate our shachain producer
 	// and from that generate the per-commitment point.
 	shaChainRoot := shachain.NewRevocationProducer(revRoot)
-	taprootShaChainRoot, err := deriveMusig2Shachain(shaChainRoot)
+	taprootShaChainRoot, err := channeldb.DeriveMusig2Shachain(shaChainRoot)
 	if err != nil {
 		return nil, nil, err
 	}
